+++ scripts/public-additional-scripts.js
@@ -4,3 +4,4 @@ import './public/dom-loader';
 import './public/driver-category-query';
 import './public/landing-banners';
 import './public/search';
+import './public/side-scroll-generator';